From dir to single gpg key Now it is not necessary to mount a directory in order to get the single gpg key, it now mounts only the public.key file and from it gets the fingerprint so it is easier to manage.

+++ entrypoint.sh
@@ -5,43 +5,40 @@ set -e
 CMD="/app/tbackbot.py"
 
 # Check if variables are set
-if [[ -n "$TG_BOT_TOKEN" ]] && [[ -n "$TG_API_ID" ]] && [[ -n "$TG_API_HASH" ]] && [[ -n "$TG_CHAT_ID" ]] && [[ -n "$GPG_KEY_FP" ]] && [[ -n "$TAR_NAME" ]]; then
+if [ -n "$TG_BOT_TOKEN" ] && [ -n "$TG_API_ID" ] && [ -n "$TG_API_HASH" ] && [ -n "$TG_CHAT_ID" ] && [ -n "$TAR_NAME" ]; then
     echo "Variables set"
 else
     echo "Missing variables"
     exit 1
 fi
 
-
-# If keys dir is not empty
-if [[ -n "$(ls -A /keys)" ]]; then
-    echo '
--------------------------------------
-|         Starting container        |
--------------------------------------
+echo '
+-----------------------------
+|     Starting container    |
+-----------------------------
 '
+
+# If keys exists
+if [[ -e /public.key ]] ; then
     # Set timezone
     rm /etc/localtime
     ln -s /usr/share/zoneinfo/${TZ} /etc/localtime
-    # Import keys
-    for file in $(ls /keys)
-    do
-	gpg --import /keys/${file}
-    done
+    # GPG import key and get the fingerprint
+    gpg --import /public.key
+    export GPG_KEY_FP=$(gpg --fingerprint --with-colons | grep fpr:: | sed 's/fpr:://g' | sed -r 's/://g' | head -n 1)
 else
-    # No keys, tell user how to create them
     echo ' 
-You are missing the gpg public keys
+You are missing the gpg public key, to
+export it:
 
-1. List keys
-
+1. First list your keys
     gpg --list-keys
 
-2. Export public key
-
+2. Export the public key to a file
     gpg --export -a <key-fingerprint> > public.key
 
-You now can save the public.key file in the 'keys' folder.
+You now can save the public.key and mount
+it in /public.key inside the container.
 '
     CMD="echo 'Exiting container ...'"
     echo
